add update tokens for removed devices

+++ src/controllers/devicesController.ts
@@ -1,10 +1,11 @@
 import {NextFunction, Request, Response} from 'express';
-import {deviceCollection} from "../db/mongo-db";
+import {deviceCollection, tokenCollection} from "../db/mongo-db";
 import {WithId} from "mongodb";
 import {IDevice} from "../types/devices.interface";
 import {tokenService} from "../services/token.service";
 import {ApiError} from "../exceptions/api.error";
 import {v4 as uuid} from 'uuid';
+import {tokensRepository} from "../repositories/tokensRepository";
 
 
 export const getDevicesController = async (req: Request<any, any, any, any>, res: Response, next: NextFunction) => {
@@ -44,6 +45,10 @@ export const deleteDeviceByIdController = async (req: Request, res: Response, ne
         if (!validateToken) {
             return next(ApiError.UnauthorizedError())
         }
+        const updateTokenInfo = await tokensRepository.updateTokenForActivate(token)
+        if (!updateTokenInfo) {
+            return  next(ApiError.UnauthorizedError())
+        }
         await deviceCollection.deleteOne({deviceId: req.params.id})
         res.status(204).send('Удалено');
     } catch (e) {
@@ -59,6 +64,7 @@ export const deleteAllDevicesExceptCurrentController = async (req: Request, res:
             return next(ApiError.UnauthorizedError())
         }
         await deviceCollection.deleteMany({deviceId: {$ne: validateToken.deviceId}})
+        await tokenCollection.updateMany({refreshToken: token}, {$set: {blackList: true}})
         res.status(204).send('Удалено');
     } catch (e) {
         res.status(500).send(e)
